initial commit
This project is an implementation of RSA for CYS 492-01: Cryptographic
Systems, Spring 2021. I chose Python for this implementation because it
handles large numbers well and has a variety of helpful libraries to
extend its functionality without compromising readability. This project
was completed with the help of Mack Gromadski, my project partner.