Added prime generation
Added the ability to find the next n primes after a number, and the
ability to find n primes within a certain range. Demonstrated these
capabilities in rsa.py. Also deleted the gcd method from prime.py for
the time being. Will add it back in for Fermat's little theorem.