Implemented encryption and decryption
Encryption/decryption with Euclid's algorithm for multiplicative
inverses is now working! It currently only works for given values of p
and q, and there are definitely things to be improved upon (for
instance, storing ciphertext as a string).

TODO: combine euclid and modInv methods in RSA.
TODO: make rsa.py into a class and use main.py to access it with
stdin/stdout and file inputs?