Added simple Diffie-Hellman code
dh.py includes the basic maths required for a Diffie-Hellman key
exchange. Sending/receiving the shared secret(s) is not going to be
easy, but it can be done. TODO: figure out how to integrate all this
with RSA for the protocol implementation.

Sidenote: use public g and p as nonce?